+++ R/dateRange.R
@@ -83,6 +83,24 @@
 #' this function), which will force \code{POSIXct} inputs into a new timezone,
 #' altering the physical moment of time the input represents.
 #'
+#' @section Parameter precedence:
+#' It is possible to supply input paramters that are in conflict. For example:
+#'
+#' \code{dateRange("2019-01-01", "2019-01-08", days = 3, timezone = "UTC")}
+#'
+#' The \code{startdate} and \code{enddate} parameters would imply a 7-day range
+#' which is in conflict with \code{days = 3}. The following rules resolve
+#' conflicts of this nature:
+#'
+#' \enumerate{
+#' \item{When \code{startdate} and \code{enddate} are both specified, the
+#' \code{days} parameter is ignored.}
+#' \item{When \code{startdate} is missing, the first returned time will depend
+#' on the combination of \code{enddate}, \code{days} and \code{ceilingEnd}.}
+#' \item{When \code{enddate} is missing, \code{ceillingEnd} is ignored and the
+#' second returned time depends on \code{days}.}
+#' }
+#'
 #' @export
 #'
 #' @examples
@@ -155,6 +173,13 @@ dateRange <- function(
       timeInput[2] %>%
       lubridate::floor_date(unit = "day")
 
+    # Ignore "days" parameter
+
+    # Adjust for ceilingEnd
+    if ( ceilingEnd ) {
+      endtime <- endtime + lubridate::ddays(1)
+    }
+
     endtime <- endtime - endUnitAdjust
 
   } else if ( !is.null(startdate) && is.null(enddate) ) {
@@ -166,24 +191,27 @@ dateRange <- function(
       parseDatetime(timezone = timezone) %>%
       lubridate::floor_date(unit = "day")
 
-    endtime <- starttime + lubridate::days(days) - endUnitAdjust
+    endtime <- starttime + lubridate::days(days)
 
+    # Ignore "ceilingEnd" parameter
+
+    endtime <- endtime - endUnitAdjust
 
   } else if ( is.null(startdate) && !is.null(enddate) ) {
 
     # ** Missing startdate: use (enddate - days), enddate ----------------------
-
-    ## NOTE:
-    #  Don't account for end unit adjustments until after calculating the start
-    #  time from the end time, in order to make the math for subtracting the
-    #  correct amount of time from the end time simpler.
 
     endtime <-
       enddate %>%
       parseDatetime(timezone = timezone) %>%
       lubridate::floor_date(unit = "day")
 
+    if ( ceilingEnd ) {
+      endtime <- endtime + lubridate::ddays(1)
+    }
+
     starttime <- endtime - lubridate::days(days)
+
     endtime <- endtime - endUnitAdjust
 
 
@@ -195,14 +223,14 @@ dateRange <- function(
       lubridate::now(tzone = timezone) %>%
       lubridate::floor_date(unit = "day")
 
+    if ( ceilingEnd ) {
+      endtime <- endtime + lubridate::ddays(1)
+    }
+
     starttime <- endtime - lubridate::days(days)
+
     endtime <- endtime - endUnitAdjust
 
-  }
-
-  # Adjust for ceilingEnd
-  if ( ceilingEnd ) {
-    endtime <- endtime + lubridate::ddays(1)
   }